Rename some files
treeDataProvider/ -> tree/
AzureActionHandler.ts -> registerCommand.ts & registerEvent.ts
