+++ ui/src/index.ts
@@ -5,7 +5,6 @@
 
 export { createAzExtOutputChannel } from './AzExtOutputChannel';
 export * from './AzExtTreeFileSystem';
-export * from './AzureActionHandler';
 export * from './AzureUserInput';
 export * from './BaseEditor';
 export * from './callWithMaskHandling';
@@ -20,15 +19,17 @@ export { registerUIExtensionVariables } from './extensionVariables';
 export * from './openInPortal';
 export * from './openReadOnlyContent';
 export * from './parseError';
+export * from './registerCommand';
+export * from './registerEvent';
 export { registerReportIssueCommand } from './registerReportIssueCommand';
-export * from './treeDataProvider/AzExtParentTreeItem';
-export * from './treeDataProvider/AzExtTreeDataProvider';
-export * from './treeDataProvider/AzExtTreeItem';
-export * from './treeDataProvider/AzureAccountTreeItemBase';
-export * from './treeDataProvider/AzureParentTreeItem';
-export * from './treeDataProvider/AzureTreeItem';
-export * from './treeDataProvider/GenericTreeItem';
-export * from './treeDataProvider/SubscriptionTreeItemBase';
+export * from './tree/AzExtParentTreeItem';
+export * from './tree/AzExtTreeDataProvider';
+export * from './tree/AzExtTreeItem';
+export * from './tree/AzureAccountTreeItemBase';
+export * from './tree/AzureParentTreeItem';
+export * from './tree/AzureTreeItem';
+export * from './tree/GenericTreeItem';
+export * from './tree/SubscriptionTreeItemBase';
 export * from './wizard/AzureNameStep';
 export * from './wizard/AzureWizard';
 export * from './wizard/AzureWizardExecuteStep';
