+++ ui/src/registerEvent.ts
@@ -1,0 +1,20 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { Event } from 'vscode';
+import * as types from '../index';
+import { callWithTelemetryAndErrorHandling } from './callWithTelemetryAndErrorHandling';
+import { ext } from './extensionVariables';
+
+export function registerEvent<T>(eventId: string, event: Event<T>, callback: (context: types.IActionContext, ...args: any[]) => any): void {
+    ext.context.subscriptions.push(event(async (...args: any[]): Promise<any> => {
+        return await callWithTelemetryAndErrorHandling(
+            eventId,
+            (context: types.IActionContext) => {
+                return callback(context, ...args);
+            }
+        );
+    }));
+}
