+++ ui/src/tree/AzureAccountTreeItemBase.ts
@@ -7,9 +7,9 @@ import * as semver from 'semver';
 import { commands, Disposable, Extension, extensions, MessageItem, ProgressLocation, window } from 'vscode';
 import * as types from '../../index';
 import { AzureAccount, AzureLoginStatus, AzureResourceFilter } from '../azure-account.api';
-import { registerEvent } from '../AzureActionHandler';
 import { UserCancelledError } from '../errors';
 import { localize } from '../localize';
+import { registerEvent } from '../registerCommand';
 import { nonNullProp, nonNullValue } from '../utils/nonNull';
 import { AzureWizardPromptStep } from '../wizard/AzureWizardPromptStep';
 import { AzExtParentTreeItem } from './AzExtParentTreeItem';
