+++ ui/src/utils/treeUtils.ts
@@ -3,8 +3,8 @@
 *  Licensed under the MIT License. See License.txt in the project root for license information.
 *--------------------------------------------------------------------------------------------*/
 
-import { AzExtTreeItem } from '../treeDataProvider/AzExtTreeItem';
-import { SubscriptionTreeItemBase } from '../treeDataProvider/SubscriptionTreeItemBase';
+import { AzExtTreeItem } from '../tree/AzExtTreeItem';
+import { SubscriptionTreeItemBase } from '../tree/SubscriptionTreeItemBase';
 
 // tslint:disable-next-line:export-name
 export function findSubscriptionTreeItem(node: AzExtTreeItem): SubscriptionTreeItemBase {
